requirements.txt
mysql-connector-python==8.0.33
streamlit==1.31.0
pandas
